% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{add_html_part}
\alias{add_html_part}
\title{Add html text of table part}
\usage{
add_html_part(st, tb, rows, tag_part, hash, indent_level)
}
\arguments{
\item{st}{A \link[=styled_table]{StyledTable} class object}

\item{tb}{A text builder}

\item{rows}{The rows of the \code{st} which should be added}

\item{tag_part}{A string holding the html tag of the table part.
Usually either \code{"thead"} or \code{"tbody"}.}

\item{hash}{A string holding the (almost) unique hash value for the table.}

\item{indent_level}{An optional integer defining, the indentation level of
the resulting html code}
}
\description{
Add html text of table part
}
