% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMerge.R
\docType{methods}
\name{addMerge}
\alias{addMerge}
\alias{addMerge,StyledTable-method}
\alias{addMerge,StyledTable,character-method}
\title{Add merged region}
\usage{
addMerge(object, ...)

\S4method{addMerge}{StyledTable}(object, rows = NULL, cols = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector holding two numbers, the start row and the end row of the merge region}

\item{cols}{A vector holding two numbers, the start col and the end col of the merge region}
}
\description{
Add merged region
}
