% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment.R
\name{Alignment}
\alias{Alignment}
\title{Bug fix in xlsx package (indent != 0 is not supported)}
\usage{
Alignment(horizontal = NULL, vertical = NULL, wrapText = FALSE,
  rotation = 0, indent = 0)
}
\arguments{
\item{horizontal}{Character string for horizontal alignement}

\item{vertical}{Character string for vertical alignement}

\item{wrapText}{Logical should cell content be wrapped}

\item{rotation}{Numerical value for cell content rotation}

\item{indent}{Number of white spaces that should be used for indentation}
}
\description{
Bug fix in xlsx package (indent != 0 is not supported)
}
