% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_latex_table.R
\name{append_latex_table}
\alias{append_latex_table}
\alias{append_latex_table,character,StyledTable-method}
\title{Append the LaTeX code for a StyledTable to an existing LaTeX table code}
\usage{
append_latex_table(table_text, st)

\S4method{append_latex_table}{character,StyledTable}(table_text = "", st)
}
\arguments{
\item{table_text}{A string containing the LaTeX code of a table}

\item{st}{A \link{StyledTable} object}
}
\value{
A character string holding the latex code for the extended styled table
}
\description{
Append the LaTeX code for a StyledTable to an existing LaTeX table code
}
\seealso{
\code{\link[=write_pdf]{write_pdf()}}, \code{\link[=write_png]{write_png()}}, \code{\link[=write_excel]{write_excel()}}, \code{\link[=create_latex_table]{create_latex_table()}}, \code{\link[=create_latex_table_body]{create_latex_table_body()}}
}
