% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\docType{methods}
\name{appendLatexTable}
\alias{appendLatexTable}
\alias{appendLatexTable,character,StyledTable-method}
\title{Append the LaTeX code for a StyledTable to an existing LaTeX table code}
\usage{
appendLatexTable(tableText, object)

\S4method{appendLatexTable}{character,StyledTable}(tableText = "", object)
}
\arguments{
\item{tableText}{A string containing the LaTeX code of a table}

\item{object}{A StyledTable}
}
\description{
Append the LaTeX code for a StyledTable to an existing LaTeX table code
}
