% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{asser_color_value}
\alias{asser_color_value}
\title{Check if a variable holds a color value (using non standard evaluation)}
\usage{
asser_color_value(var)
}
\arguments{
\item{var}{Variable whose value should be checked}
}
\description{
Check if a variable holds a color value (using non standard evaluation)
}
