% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_type_}
\alias{assert_type_}
\title{Check if type is as expected (without using non standard evaluation)}
\usage{
assert_type_(var, types, var_name, allow_na = TRUE)
}
\arguments{
\item{var}{Variable whose type should be checked}

\item{types}{(vector of strings) A vector of allowed type values}

\item{var_name}{(string) Name of the variable passed in in \code{var}. This
name is used for the error message if it has the wrong type.}

\item{allow_na}{(logical) Is a missing value allowed?}
}
\description{
Check if type is as expected (without using non standard evaluation)
}
