% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_value}
\alias{assert_value}
\title{Check if a variable holds an allowed value (using non standard evaluation)}
\usage{
assert_value(var, values)
}
\arguments{
\item{var}{Variable whose value should be checked}

\item{values}{(vector) A vector of allowed values}
}
\description{
Check if a variable holds an allowed value (using non standard evaluation)
}
