% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_value_}
\alias{assert_value_}
\title{Check if a variable holds an allowed value (without non standard evaluation)}
\usage{
assert_value_(var, var_name, values)
}
\arguments{
\item{var}{Variable whose value should be checked}

\item{var_name}{name of the variable whose value should be checked}

\item{values}{(vector) A vector of allowed values}
}
\description{
Check if a variable holds an allowed value (without non standard evaluation)
}
