% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AssertChar}
\alias{AssertChar}
\title{Check if a variable is a character, if not throw an error (using non standard evaluation)}
\usage{
AssertChar(var)
}
\arguments{
\item{var}{The value that should be checked}
}
\description{
Check if a variable is a character, if not throw an error (using non standard evaluation)
}
