% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AssertLogical}
\alias{AssertLogical}
\title{Check if a variable is a logical, if not throw an error (using non standard evaluation)}
\usage{
AssertLogical(var)
}
\arguments{
\item{var}{The value that should be checked}
}
\description{
Check if a variable is a logical, if not throw an error (using non standard evaluation)
}
