% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AssertNumber}
\alias{AssertNumber}
\title{Check if a variable is a number, if not throw an error (using non standard evaluation)}
\usage{
AssertNumber(var)
}
\arguments{
\item{var}{The value that should be checked}
}
\description{
Check if a variable is a number, if not throw an error (using non standard evaluation)
}
