% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assertType}
\alias{assertType}
\title{Check if type is as expected (without using non standard evaluation)}
\usage{
assertType(var, types, varName, allowNA = TRUE)
}
\arguments{
\item{var}{Variable whose type should be checked}

\item{types}{(vector of strings) A vector of allowed type values}

\item{varName}{(string) Name of the variable passed in in \code{var}. This
name is used for the error message if it has the wrong type.}

\item{allowNA}{(logical) Is a missing value allowed?}
}
\description{
Check if type is as expected (without using non standard evaluation)
}
