% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assertValue_}
\alias{assertValue_}
\title{Check if a variable holds an allowed value (without non standard evaluation)}
\usage{
assertValue_(var, varName, values)
}
\arguments{
\item{var}{Variable whose value should be checked}

\item{varName}{name of the variable whose value should be checked}

\item{values}{(vector) A vector of allowed values}
}
\description{
Check if a variable holds an allowed value (without non standard evaluation)
}
