% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{between_vec}
\alias{between_vec}
\title{Vectorized version of 'between'}
\usage{
between_vec(x, vec)
}
\arguments{
\item{x}{A numeric value that should be between two other values}

\item{vec}{A vector holding the upper and lower bound values}
}
\description{
Vectorized version of 'between'
}
