% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{betweenVec}
\alias{betweenVec}
\title{Vectorized version of 'between'}
\usage{
betweenVec(x, vec)
}
\arguments{
\item{x}{A numeric value that should be between two other values}

\item{vec}{A vector holding the upper and lower bound values}
}
\description{
Vectorized version of 'between'
}
