% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_hex_color}
\alias{calc_hex_color}
\title{Calculate the hex color value from a color name}
\usage{
calc_hex_color(col)
}
\arguments{
\item{col}{A string holding a color name (e.g.: "red") or a color value (e.g.: "#F00000")}
}
\value{
The hex color value as string without hashtag (e.g.: "F00000")
}
\description{
This little function takes a color name (e.g.: "red") and calculates the corresponding hex color value.
}
