% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{rowheader_id}
\alias{rowheader_id}
\alias{subheading_id}
\alias{colheader_id}
\title{Define cell ids for accessibility}
\usage{
rowheader_id(i, j, hash)

subheading_id(i, hash)

colheader_id(i, j, hash)
}
\arguments{
\item{i}{Row number}

\item{j}{Column number}

\item{hash}{A string holding the (almost) unique hash value for the table.}
}
\value{
A string holding the cell-id
}
\description{
Define cell ids for accessibility
}
