% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_cell.R
\name{check_validity}
\alias{check_validity}
\title{Check Validity of StyledCell class sc and return an error description, if there is an error}
\usage{
check_validity(sc, style_name)
}
\arguments{
\item{sc}{StyledCell object}

\item{style_name}{A character holding the name of the style that should be validated}
}
\description{
Check Validity of StyledCell class sc and return an error description, if there is an error
}
