% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\name{checkValidity}
\alias{checkValidity}
\title{Check Validity of STCellStyle class object and return an error description, if there is an error}
\usage{
checkValidity(object, styleName)
}
\arguments{
\item{object}{STCellStyle object}

\item{styleName}{A character holding the name of the style that should be validated}
}
\description{
Check Validity of STCellStyle class object and return an error description, if there is an error
}
