% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{concat_functions}
\alias{concat_functions}
\title{Concatenate functions}
\usage{
concat_functions(...)
}
\arguments{
\item{...}{Several functions that should be concatenated}
}
\description{
Concatenate functions
}
