% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{concatFunctions}
\alias{concatFunctions}
\title{Concatenate functions}
\usage{
concatFunctions(...)
}
\arguments{
\item{...}{Several functions that should be concatenated}
}
\description{
Concatenate functions
}
