% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{create_html_tag}
\alias{create_html_tag}
\title{Create the text for an html tag opening}
\usage{
create_html_tag(tag, properties, close = FALSE)
}
\arguments{
\item{tag}{String holding the tag name}

\item{properties}{A character vector holding the tag properties. Each
entry could be created by \code{\link[=create_html_tag_property]{create_html_tag_property()}}.}

\item{close}{An optional logical flag, defining if the html tag should be
immediately be closed (eg. \verb{<tag/>}).}
}
\value{
A string holding the text for the html opening tag.
}
\description{
Create the text for an html tag opening
}
