% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{create_html_tag_property}
\alias{create_html_tag_property}
\title{Create a string defining a property of an html tag}
\usage{
create_html_tag_property(name, values = NULL)
}
\arguments{
\item{name}{Name of the property}

\item{values}{A character vector holding a set of values which should be assigned
to the property.}
}
\value{
A string holding the text for the html tag property.
}
\description{
Create a string defining a property of an html tag
}
