% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{create_html_tag_style}
\alias{create_html_tag_style}
\title{Create a string defining the style property of an html tag}
\usage{
create_html_tag_style(styles, values = NULL)
}
\arguments{
\item{styles}{An character vector holding the names of the style properties for an html dom element.}

\item{values}{A vector of same length as \code{styles} holding the corresponding
values.}
}
\value{
A string holding the text for the html tag styling property
}
\description{
Create a string defining the style property of an html tag
}
