% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_latex_table.R
\docType{methods}
\name{create_latex_table}
\alias{create_latex_table}
\alias{create_latex_table,StyledTable-method}
\title{Generate the LaTeX code for a StyledTable}
\usage{
create_latex_table(st)

\S4method{create_latex_table}{StyledTable}(st)
}
\arguments{
\item{st}{A \link{StyledTable} object}
}
\value{
A character string holding the latex code for the styled table
}
\description{
Generate the LaTeX code for a StyledTable
}
\seealso{
\code{\link[=write_pdf]{write_pdf()}}, \code{\link[=write_png]{write_png()}}, \code{\link[=write_excel]{write_excel()}}, \code{\link[=append_latex_table]{append_latex_table()}}, \code{\link[=create_latex_table_body]{create_latex_table_body()}}
}
