% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\docType{methods}
\name{createLatexTable}
\alias{createLatexTable}
\alias{createLatexTable,StyledTable-method}
\title{Generate the LaTeX code for a StyledTable}
\usage{
createLatexTable(object)

\S4method{createLatexTable}{StyledTable}(object)
}
\arguments{
\item{object}{A StyledTable}
}
\description{
Generate the LaTeX code for a StyledTable
}
