% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\docType{methods}
\name{createLatexTableBody}
\alias{createLatexTableBody}
\alias{createLatexTableBody,StyledTable-method}
\title{Generate the LaTeX code for a StyledTable without begin{tabular} and end{tabular}}
\usage{
createLatexTableBody(object)

\S4method{createLatexTableBody}{StyledTable}(object)
}
\arguments{
\item{object}{A StyledTable}
}
\description{
Generate the LaTeX code for a StyledTable without begin{tabular} and end{tabular}
}
