% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{doCallWithoutMissing}
\alias{doCallWithoutMissing}
\title{Call function with arguments and remove all empty arguments (NULL or length == 0)}
\usage{
doCallWithoutMissing(fn, args, skipIfAllMissing = TRUE)
}
\arguments{
\item{fn}{(function) Function that should be invoked}

\item{args}{(named list) List of arguments (\code{NULL} arguments will be removed)}

\item{skipIfAllMissing}{(logical) If \code{skipIfAllMissing = TRUE} and all arguments in \code{args} are \code{NULL}, then the function \code{fn} is not invoked and \code{NULL} is returned instead.}
}
\description{
Invokes a function with a list of arguments, but removes all \code{NULL} arguments first
}
