% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{FILL_STYLES_}
\alias{FILL_STYLES_}
\title{Import the non-exported FILL_STYLES_ from the xlsx package namespace}
\format{
An object of class \code{numeric} of length 19.
}
\usage{
FILL_STYLES_
}
\description{
Import the non-exported FILL_STYLES_ from the xlsx package namespace
}
\keyword{datasets}
