% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stat.R
\docType{methods}
\name{format_stat_header}
\alias{format_stat_header}
\alias{format_stat_header,StyledTable-method}
\title{Method format_stat_header.}
\usage{
format_stat_header(st, ...)

\S4method{format_stat_header}{StyledTable}(st, row_id = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{...}{Various arguments}

\item{row_id}{A vector of row numbers. If omitted, the formatting will be applied to all rows}
}
\value{
The modified \link{StyledTable} object
}
\description{
Method format_stat_header.
}
