% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stat.R
\docType{methods}
\name{format_stat_relative}
\alias{format_stat_relative}
\alias{format_stat_relative,StyledTable-method}
\title{Method format_stat_relative}
\usage{
format_stat_relative(st, ...)

\S4method{format_stat_relative}{StyledTable}(st, row_id = NULL,
  col_id = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{...}{Various arguments}

\item{row_id}{A vector of row numbers}

\item{col_id}{A vector of col numbers}
}
\value{
The modified \link{StyledTable} object
}
\description{
Method format_stat_relative
}
