% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStat.R
\docType{methods}
\name{formatStatAbsolute}
\alias{formatStatAbsolute}
\alias{formatStatAbsolute,StyledTable-method}
\title{Method formatStatAbsolute}
\usage{
formatStatAbsolute(object, ...)

\S4method{formatStatAbsolute}{StyledTable}(object, rows = NULL, cols = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers}

\item{cols}{A vector of col numbers}
}
\description{
Method formatStatAbsolute
}
