% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStat.R
\docType{methods}
\name{formatStatBody}
\alias{formatStatBody}
\alias{formatStatBody,StyledTable-method}
\title{Method formatStatBody.}
\usage{
formatStatBody(object, ...)

\S4method{formatStatBody}{StyledTable}(object, rows = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers (N is substituted as \code{nRows(object)})}
}
\description{
Method formatStatBody.
}
