% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStat.R
\docType{methods}
\name{formatStatHeader}
\alias{formatStatHeader}
\alias{formatStatHeader,StyledTable-method}
\title{Method formatStatHeader.}
\usage{
formatStatHeader(object, ...)

\S4method{formatStatHeader}{StyledTable}(object, rows = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers (N is substituted as \code{nRows(object)})}
}
\description{
Method formatStatHeader.
}
