% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStat.R
\docType{methods}
\name{formatStatRelative}
\alias{formatStatRelative}
\alias{formatStatRelative,StyledTable-method}
\title{Method formatStatRelative}
\usage{
formatStatRelative(object, ...)

\S4method{formatStatRelative}{StyledTable}(object, rows = NULL, cols = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers}

\item{cols}{A vector of col numbers}
}
\description{
Method formatStatRelative
}
