% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStat.R
\docType{methods}
\name{formatStatSubHeading}
\alias{formatStatSubHeading}
\alias{formatStatSubHeading,StyledTable-method}
\title{Method formatStatSubHeading}
\usage{
formatStatSubHeading(object, ...)

\S4method{formatStatSubHeading}{StyledTable}(object, rows = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers}
}
\description{
Method formatStatSubHeading
}
