% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_cell_styles.R
\name{get_xlsx_border}
\alias{get_xlsx_border}
\alias{get_xlsx_border,StyledCell-method}
\title{Method get_xlsx_border}
\usage{
get_xlsx_border(sc)

\S4method{get_xlsx_border}{StyledCell}(sc)
}
\arguments{
\item{sc}{A StyledCell object}
}
\description{
Method get_xlsx_border
}
