% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_cell_styles.R
\name{get_xlsx_cell_protection}
\alias{get_xlsx_cell_protection}
\alias{get_xlsx_cell_protection,StyledCell-method}
\title{Method get_xlsx_cell_protection}
\usage{
get_xlsx_cell_protection(sc)

\S4method{get_xlsx_cell_protection}{StyledCell}(sc)
}
\arguments{
\item{sc}{A StyledCell object}
}
\description{
Method get_xlsx_cell_protection
}
