% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_cell_styles.R
\name{get_xlsx_cell_style}
\alias{get_xlsx_cell_style}
\alias{get_xlsx_cell_style,ANY,StyledCell-method}
\title{Method get_xlsx_cell_style}
\usage{
get_xlsx_cell_style(wb, sc)

\S4method{get_xlsx_cell_style}{ANY,StyledCell}(wb, sc)
}
\arguments{
\item{wb}{An xlsx workbook object}

\item{sc}{A StyledCell object}
}
\description{
Method get_xlsx_cell_style
}
