% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_cell_styles.R
\docType{methods}
\name{get_xlsx_data_format}
\alias{get_xlsx_data_format}
\alias{get_xlsx_data_format,StyledCell-method}
\title{Method get_xlsx_data_format}
\usage{
get_xlsx_data_format(sc)

\S4method{get_xlsx_data_format}{StyledCell}(sc)
}
\arguments{
\item{sc}{A StyledCell object}
}
\description{
Method get_xlsx_data_format
}
