% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_cell_styles.R
\docType{methods}
\name{get_xlsx_fill}
\alias{get_xlsx_fill}
\alias{get_xlsx_fill,StyledCell-method}
\title{Method get_xlsx_fill}
\usage{
get_xlsx_fill(sc)

\S4method{get_xlsx_fill}{StyledCell}(sc)
}
\arguments{
\item{sc}{A StyledCell object}
}
\description{
Method get_xlsx_fill
}
