% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getAlignmentCellStyle}
\alias{getAlignmentCellStyle}
\alias{getAlignmentCellStyle,Style-method}
\title{Method getAlignmentCellStyle}
\usage{
getAlignmentCellStyle(object)

\S4method{getAlignmentCellStyle}{Style}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getAlignmentCellStyle
}
