% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getBorderCellStyle}
\alias{getBorderCellStyle}
\alias{getBorderCellStyle,Style-method}
\title{Method getBorderCellStyle}
\usage{
getBorderCellStyle(object)

\S4method{getBorderCellStyle}{Style}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getBorderCellStyle
}
