% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getCellProtectionCellStyle}
\alias{getCellProtectionCellStyle}
\alias{getCellProtectionCellStyle,STCellStyle-method}
\title{Method getCellProtectionCellStyle}
\usage{
getCellProtectionCellStyle(object)

\S4method{getCellProtectionCellStyle}{STCellStyle}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getCellProtectionCellStyle
}
