% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getCellStyle}
\alias{getCellStyle}
\alias{getCellStyle,ANY,Style-method}
\title{Method getCellStyle}
\usage{
getCellStyle(wb, object)

\S4method{getCellStyle}{ANY,Style}(wb, object)
}
\arguments{
\item{wb}{An xlsx workbook object}

\item{object}{A Style object}
}
\description{
Method getCellStyle
}
