% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getDataFormatCellStyle}
\alias{getDataFormatCellStyle}
\alias{getDataFormatCellStyle,Style-method}
\title{Method getDataFormatCellStyle}
\usage{
getDataFormatCellStyle(object)

\S4method{getDataFormatCellStyle}{Style}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getDataFormatCellStyle
}
