% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getFillCellStyle}
\alias{getFillCellStyle}
\alias{getFillCellStyle,Style-method}
\title{Method getFillCellStyle}
\usage{
getFillCellStyle(object)

\S4method{getFillCellStyle}{Style}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getFillCellStyle
}
