% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getFontCellStyle}
\alias{getFontCellStyle}
\alias{getFontCellStyle,ANY,Style-method}
\alias{getFontCellStyle,Style-method}
\title{Method getFontCellStyle}
\usage{
getFontCellStyle(wb, object)

\S4method{getFontCellStyle}{ANY,Style}(wb, object)
}
\arguments{
\item{wb}{An xlsx workbook object}

\item{object}{A Style object}
}
\description{
Method getFontCellStyle
}
