% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getSTCellStyle}
\alias{getSTCellStyle}
\alias{getSTCellStyle,STCellStyle-method}
\alias{getSTCellStyle,MissingOrNull-method}
\title{Method getSTCellStyle}
\usage{
getSTCellStyle(object, ...)

\S4method{getSTCellStyle}{STCellStyle}(object, styleName)

\S4method{getSTCellStyle}{MissingOrNull}(object, styleName)
}
\arguments{
\item{object}{A STCellStyle object}

\item{...}{Various arguments}

\item{styleName}{The name of the style setting that should be retrieved}
}
\description{
Method getSTCellStyle
}
