% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getStyle}
\alias{getStyle}
\alias{getStyle,Style-method}
\alias{getStyle,MissingOrNull-method}
\title{Method getStyle}
\usage{
getStyle(object, ...)

\S4method{getStyle}{Style}(object, styleName)

\S4method{getStyle}{MissingOrNull}(object, styleName)
}
\arguments{
\item{object}{A Style object}

\item{...}{Various arguments}

\item{styleName}{The name of the style setting that should be retrieved}
}
\description{
Method getStyle
}
