% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_cell.R
\docType{methods}
\name{getStyledCell}
\alias{getStyledCell}
\alias{getStyledCell,StyledCell-method}
\alias{getStyledCell,MissingOrNull-method}
\title{Method getStyledCell}
\usage{
getStyledCell(sc, ...)

\S4method{getStyledCell}{StyledCell}(sc, style_name)

\S4method{getStyledCell}{MissingOrNull}(sc, style_name)
}
\arguments{
\item{sc}{A StyledCell object}

\item{...}{Various arguments}

\item{style_name}{The name of the style setting that should be retrieved}
}
\description{
Method getStyledCell
}
