% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getXlsxAlignmentCellStyle}
\alias{getXlsxAlignmentCellStyle}
\alias{getXlsxAlignmentCellStyle,STCellStyle-method}
\title{Method getXlsxAlignmentCellStyle}
\usage{
getXlsxAlignmentCellStyle(object)

\S4method{getXlsxAlignmentCellStyle}{STCellStyle}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getXlsxAlignmentCellStyle
}
