% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getXlsxBorderCellStyle}
\alias{getXlsxBorderCellStyle}
\alias{getXlsxBorderCellStyle,STCellStyle-method}
\title{Method getXlsxBorderCellStyle}
\usage{
getXlsxBorderCellStyle(object)

\S4method{getXlsxBorderCellStyle}{STCellStyle}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getXlsxBorderCellStyle
}
