% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getXlsxCellStyle}
\alias{getXlsxCellStyle}
\alias{getXlsxCellStyle,ANY,STCellStyle-method}
\title{Method getXlsxCellStyle}
\usage{
getXlsxCellStyle(wb, object)

\S4method{getXlsxCellStyle}{ANY,STCellStyle}(wb, object)
}
\arguments{
\item{wb}{An xlsx workbook object}

\item{object}{A Style object}
}
\description{
Method getXlsxCellStyle
}
