% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getXlsxDataFormatCellStyle}
\alias{getXlsxDataFormatCellStyle}
\alias{getXlsxDataFormatCellStyle,STCellStyle-method}
\title{Method getXlsxDataFormatCellStyle}
\usage{
getXlsxDataFormatCellStyle(object)

\S4method{getXlsxDataFormatCellStyle}{STCellStyle}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getXlsxDataFormatCellStyle
}
