% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getXlsxFillCellStyle}
\alias{getXlsxFillCellStyle}
\alias{getXlsxFillCellStyle,STCellStyle-method}
\title{Method getXlsxFillCellStyle}
\usage{
getXlsxFillCellStyle(object)

\S4method{getXlsxFillCellStyle}{STCellStyle}(object)
}
\arguments{
\item{object}{A Style object}
}
\description{
Method getXlsxFillCellStyle
}
