% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{getXlsxFontCellStyle}
\alias{getXlsxFontCellStyle}
\alias{getXlsxFontCellStyle,ANY,STCellStyle-method}
\alias{getXlsxFontCellStyle,STCellStyle-method}
\title{Method getXlsxFontCellStyle}
\usage{
getXlsxFontCellStyle(wb, object)

\S4method{getXlsxFontCellStyle}{ANY,STCellStyle}(wb, object)
}
\arguments{
\item{wb}{An xlsx workbook object}

\item{object}{A Style object}
}
\description{
Method getXlsxFontCellStyle
}
