% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{lua_remove_filter}
\alias{lua_remove_filter}
\title{Pandoc argument for adding the \strong{lua filter} \code{lua_remove.lua}}
\usage{
lua_remove_filter()
}
\description{
Knitting an \strong{RMD} file containing \code{styledTables} to \strong{HTML}, then
the chunk option \code{tab.add_dummy_table} must be set to \code{TRUE}, since
the default \strong{lua filter} \code{loft.lua} (see \href{https://github.com/rstudio/pagedown/blob/main/inst/resources/lua/loft.lua}{pagedown}) for adding the \strong{list of tables} at the report
beginning does not recognize \verb{<table>} tags. Therefore, a dummy table
is inserted before the current table. The \strong{lua filter} \code{lua_remove.lua}
should be applied after \code{loft.lua}. This functions returns character vector
defining the required pandoc argument in order to add \code{lua_remove.lua} to
the list of lua filters of pandoc.
}
