% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cells.R
\docType{methods}
\name{merge_equal_cells}
\alias{merge_equal_cells}
\alias{merge_equal_cells,StyledTable-method}
\alias{merge_equal_cells,StyledTable,character-method}
\title{Merge all cells of \link{StyledTable} object whose cell values are equal}
\usage{
merge_equal_cells(st, ...)

\S4method{merge_equal_cells}{StyledTable}(st, row_id = NULL,
  col_id = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{...}{Various arguments}

\item{row_id}{A vector holding the row numbers the start row and the end row of the merge region}

\item{col_id}{A vector holding two numbers, the start col and the end col of the merge region}
}
\value{
The modified \link{StyledTable} object
}
\description{
Merges all cells in a \link{StyledTable} object that hold the same value.
This function can produce multiple merged regions at once.
}
\seealso{
\code{\link[=merge_cells]{merge_cells()}}

\code{\link[=remove_col]{remove_col()}}, \code{\link[=remove_row]{remove_row()}}
}
