% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMerge.R
\docType{methods}
\name{mergeEqualCells}
\alias{mergeEqualCells}
\alias{mergeEqualCells,StyledTable-method}
\alias{mergeEqualCells,StyledTable,character-method}
\title{Merge all cells whose cell values are equal}
\usage{
mergeEqualCells(object, ...)

\S4method{mergeEqualCells}{StyledTable}(object, rows = NULL,
  cols = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{rows}{A vector holding the row numbers the start row and the end row of the merge region}

\item{cols}{A vector holding two numbers, the start col and the end col of the merge region}
}
\description{
Merges all styled table cells that have the same value. Can generate several
merged regions at once.
}
