% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_cell.R
\docType{class}
\name{MissingOrNull-class}
\alias{MissingOrNull-class}
\title{S4 helper class that summarizes NULL and missing values}
\description{
This class is used as signature class for missing or NULL values
}
