% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_row_col.R
\name{remove_col}
\alias{remove_col}
\alias{remove_col,StyledTable-method}
\alias{removeColuqns,StyledTable,ANY-method}
\title{Remove columns from a \link{StyledTable} object}
\usage{
remove_col(st, col_id)

\S4method{remove_col}{StyledTable}(st, col_id)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{col_id}{A vector of col numbers which should be removed from the \link{StyledTable} object.}
}
\value{
The modified \link{StyledTable} object
}
\description{
Remove columns from a \link{StyledTable} object
}
