% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_missing}
\alias{remove_missing}
\title{Remove empty elements from a list}
\usage{
remove_missing(li)
}
\arguments{
\item{li}{A list}
}
\description{
Remove all elements who are totally nested NULLs or elements of length == 0) in the first level of lists
}
