% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_row_col.R
\docType{methods}
\name{remove_row}
\alias{remove_row}
\alias{remove_row,StyledTable-method}
\alias{remove_row,StyledTable,ANY-method}
\title{Remove rows from a \link{StyledTable} object}
\usage{
remove_row(st, row_id)

\S4method{remove_row}{StyledTable}(st, row_id)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{row_id}{A vector of row numbers which should be removed from the \link{StyledTable} object.}
}
\value{
The modified \link{StyledTable} object
}
\description{
Remove rows from a \link{StyledTable} object
}
