% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRowsColumns.R
\docType{methods}
\name{removeColumns}
\alias{removeColumns}
\alias{removeColumns,StyledTable-method}
\alias{removeColuqns,StyledTable,ANY-method}
\title{Remove columns from a styledTable}
\usage{
removeColumns(object, cols)

\S4method{removeColumns}{StyledTable}(object, cols)
}
\arguments{
\item{object}{A StyledTable}

\item{cols}{A vector of col numbers which should be removed from the styled table.}
}
\description{
Remove columns from a styledTable
}
