% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRowsColumns.R
\docType{methods}
\name{removeRows}
\alias{removeRows}
\alias{removeRows,StyledTable-method}
\alias{removeRows,StyledTable,ANY-method}
\title{Remove rows from a styledTable}
\usage{
removeRows(object, rows)

\S4method{removeRows}{StyledTable}(object, rows)
}
\arguments{
\item{object}{A StyledTable}

\item{rows}{A vector of row numbers which should be removed from the styled table.}
}
\description{
Remove rows from a styledTable
}
