% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq.save}
\alias{seq.save}
\title{Improved \code{seq} command that returns an empty vector if lower > upper boundary}
\usage{
\method{seq}{save}(from, to)
}
\arguments{
\item{from}{First element in the sequence}

\item{to}{Last element in the sequence}
}
\value{
The sequence
}
\description{
Enhance the \code{seq} such that it can handle the situation if \code{from > to}
}
