% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq_vec}
\alias{seq_vec}
\title{Improved \code{seq} command which allows an two element vector holding the sequence boundaries}
\usage{
seq_vec(bLU)
}
\arguments{
\item{bLU}{A two element vector. The first element is the lower boundary the second element the upper boundarie. If the lower boundary is greater than the upper boundary, then an empty vector is returned.}
}
\value{
The sequence
}
\description{
The boundaries for the sequence command are given as two element vector and passed to the \code{seq_save} function.
}
