% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_height_width.R
\name{set_excel_col_width}
\alias{set_excel_col_width}
\alias{set_excel_col_width,StyledTable,numeric-method}
\title{Method set_excel_col_width.}
\usage{
set_excel_col_width(st, value, ...)

\S4method{set_excel_col_width}{StyledTable,numeric}(st, value, col_id = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{value}{A numeric vector (length 1 or same lengt as \code{col_id}) holding the column widths}

\item{...}{Various arguments}

\item{col_id}{A vector of column numbers (N is substituted as \code{count_cols(st)})}
}
\value{
The modified \link{StyledTable} object
}
\description{
Method set_excel_col_width.
}
