% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_height_width.R
\name{set_height_width}
\alias{set_height_width}
\title{Function to set the excel_row_height/latex_padding_top/latex_padding_bottom/excel_col_width/latex_col_width}
\usage{
set_height_width(st, slot_name, id, value)
}
\arguments{
\item{st}{The \link{StyledTable} object whoose heights/widths should be changed}

\item{slot_name}{The name of the S4-class slot that should be changed (\code{excel_row_height}, \code{latex_padding_top}, \code{excel_col_width}, \code{latex_col_width})}

\item{id}{The row/column ids that should be changed (subset of 1:N where N is the number of rows/cols of the \link{StyledTable} object).}

\item{value}{A vector holding the row height/col width that should be applied}
}
\description{
Function to set the excel_row_height/latex_padding_top/latex_padding_bottom/excel_col_width/latex_col_width
}
