% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_html.R
\name{set_html_rowheader}
\alias{set_html_rowheader}
\alias{set_html_colheader}
\alias{set_html_subheading}
\alias{set_html_table_class}
\alias{remove_html_table_class}
\alias{set_html_table_id}
\alias{remove_html_table_id}
\alias{set_html_tr_class}
\alias{remove_html_tr_class}
\alias{set_html_td_class}
\alias{set_html_td_id}
\alias{remove_html_td_class}
\alias{set_html_pre_process}
\alias{add_html_stylesheet}
\alias{add_html_dependency}
\alias{apply_html_theme1}
\title{Define the HTML specific properties of a \link[=styled_table]{StyledTable} class object}
\usage{
set_html_rowheader(st, col_id)

set_html_colheader(st, row_id)

set_html_subheading(st, row_id)

set_html_table_class(st, class, replace_mode = "append")

remove_html_table_class(st, class)

set_html_table_id(st, id, replace_mode = "append")

remove_html_table_id(st, id)

set_html_tr_class(st, class, row_id = NULL, replace_mode = "append")

remove_html_tr_class(st, class, row_id = NULL)

set_html_td_class(
  st,
  class,
  row_id = NULL,
  col_id = NULL,
  replace_mode = "append"
)

set_html_td_id(st, id, row_id = NULL, col_id = NULL, replace_mode = "append")

remove_html_td_class(st, class, row_id = NULL, col_id = NULL)

set_html_pre_process(
  st,
  fn,
  row_id = NULL,
  col_id = NULL,
  replace_mode = "append"
)

add_html_stylesheet(st, file_path, name = random_name(), version = "0.0.0")

add_html_dependency(st, dep)

apply_html_theme1(st)
}
\arguments{
\item{st}{A \link[=styled_table]{StyledTable} class object}

\item{col_id}{A numeric vector holding the ids of the specified table columns.}

\item{row_id}{A numeric vector holding the ids of the specified table rows.}

\item{class}{A character vector holding the \strong{html class names},
which should be added to the specific dom element}

\item{replace_mode}{One of the following strings:
\itemize{
\item \code{"append"}: Append the value to already assigned values.
\item \code{"replace"}: Replace already assigned values by the current value.
}}

\item{id}{A character vector holding one or more \strong{html ids} which should be
assigend to the specific dom element.}

\item{fn}{A function, which takes a single argument (\code{function(x) {...}}).
If \code{\link[=create_html]{create_html()}} is called, then this function will be applied to the
value of each cell which is included in \code{col_id} and \code{row_id}.}

\item{file_path}{The path to the \strong{css} file holding the table styling.}

\item{name}{A string holding the library name (has no effect on the output).}

\item{version}{A string holding the version of the stylesheet (has no effect on the output)}

\item{dep}{An \strong{html dependecy} created by \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}}
}
\description{
The following functions allow to configure the \strong{html} output behaviour of
a \link[=styled_table]{StyledTable} class object:
\itemize{
\item \code{set_html_table_id()} or \code{remove_html_table_id()}: Assign or remove an
html \code{id} to the surrounding \verb{<table>} tag
\item \code{set_html_table_class()} or \code{remove_html_table_class()}: Append or remove an
html \code{class} name to the surrounding \verb{<table>} tag
\item \code{set_html_tr_class()} or \code{remove_html_tr_class()}: Append or remove an
html \code{class} name to the \verb{<tr>} tag of one ore more table rows
\item \code{set_html_td_class()} or \code{remove_html_td_class()}: Append or remove an
html \code{class} name to one or more table cells (\verb{<td>} or \verb{<th>} tags)
\item \code{set_html_td_id()}: Assign an html \code{id} to one ore more table cells (\verb{<td>} or \verb{<th>} tags)
\item \code{set_html_colheader()}: Define one or more table rows as header rows, holding
the headings of the columns
\item \code{set_html_rowheader()}: Define one or more table columns as header columns, holding
additional headings for the rows
\item \code{set_html_subheading()}: Define one or more table rows as subheading rows, grouping
the html table into smaller tables
\item \code{set_html_pre_process()}: Add an html specific pre-processing function to
one or more table cells. This functions will modify the cell values,
when \code{\link[=create_html]{create_html()}} is called.
\item \code{apply_html_theme1()}: Assign the html class \code{styledtable-theme1} to the \verb{<table>} tag.
This will ensure that the styling defined in
\code{styledTables/inst/stylesheets/sass_files/styledtable_theme1.sass} is used.
If you want to create your own html stylings, then you should have a look
at this sass file and create your own stylesheets (\strong{sass} or \strong{css}).
Note that you can either include your own stylesheets in the
\code{css} field in the output-format section of the \strong{yaml header} of your
\strong{rmd} file (\strong{sass} and \strong{css} files are both possible) or
you can add your stylesheet as \strong{html dependency} by calling
\code{\link[=add_html_stylesheet]{add_html_stylesheet()}}
\item \code{add_html_dependency()}: Adds an \strong{html dependency} (stylesheet, javascript etc.)
to your styledTable object. When the table is \link[knitr:knit]{printed}, then these
dependencies will automatically be included in the resulting \strong{html} header.
\item \code{add_html_stylesheet()}: A wrapper for \code{add_html_dependency()} for
adding \strong{css} (holding costom html stylings for your tables) to the
dependencies list of the \link[=styled_table]{StyledTable} class
object.
}
}
