% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_height_width.R
\name{set_html_row_height}
\alias{set_html_row_height}
\alias{set_html_row_height,StyledTable,character-method}
\alias{setHtmlRowHeigths,StyledTable,numeric-method}
\title{Method set_html_row_height.}
\usage{
set_html_row_height(st, value, ...)

\S4method{set_html_row_height}{StyledTable,character}(st, value, row_id = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{value}{A numeric vector (length 1 or same lengt as \code{row_id}) holding the row heights}

\item{...}{Various arguments}

\item{row_id}{A vector of row numbers (N is substituted as \code{count_rows(st)})}
}
\value{
The modified \link{StyledTable} object
}
\description{
Method set_html_row_height.
}
