% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_height_width.R
\name{set_latex_padding_bottom}
\alias{set_latex_padding_bottom}
\alias{set_latex_padding_bottom,StyledTable,character-method}
\title{Method set_latex_padding_bottom.}
\usage{
set_latex_padding_bottom(st, value, ...)

\S4method{set_latex_padding_bottom}{StyledTable,character}(st, value, row_id = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{value}{A character vector (length 1 or same lengt as \code{row_id}) holding the row heights}

\item{...}{Various arguments}

\item{row_id}{A vector of row numbers (N is substituted as \code{count_rows(st)})}
}
\value{
The modified \link{StyledTable} object
}
\description{
Method set_latex_padding_bottom.
}
