% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_styled_table.R
\docType{methods}
\name{set_latex_vertical_move}
\alias{set_latex_vertical_move}
\alias{set_latex_vertical_move,StyledTable-method}
\alias{set_latex_vertical_move,StyledTable,ANY-method}
\title{Set vertical movement of cell content for LaTeX table generation}
\usage{
set_latex_vertical_move(st, value, ...)

\S4method{set_latex_vertical_move}{StyledTable}(st, value, row_id = NULL,
  col_id = NULL, condition = NULL, condition_text = NULL)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{value}{The function that should be used for pre processing}

\item{...}{Various arguments}

\item{row_id}{A vector of row ids to which the change should be applied to. The variable \code{n_row} can be used in the expression to name the total number of columns (will be replaced). If the argument \code{row_id} is omitted, then the change will be applied to all rows}

\item{col_id}{A vector of column ids to which the change should be applied to. The variable \code{n_col} can be used in the expression to name the total number of columns (will be replaced). If the argument \code{col_id} is omitted, then the change will be applied to all columns}

\item{condition}{An equation (non standard evaluation) that is evaluated on the data.frame \code{data} in order to decide on which rows or cells the change should be applied. In the equation the column names of the \link{StyledTable} object should be used. Be aware that the \link{StyledTable} object columns have the names \code{X1}, ..., \code{XN}, where \code{N} is the total number of columns. If you do not want to apply the change to all columns given in the \code{col_id} argument, but only to single cells which fullfill a special condition you can use also the variable name \code{X} which is substituted by each column name separately. The arguments \code{condition_text} and \code{condition} cannot be passed at the same time.}

\item{condition_text}{An character string holding an equation that is evaluated on the data.frame \code{data} in order to decide on which rows or cells the change should be applied. In the equation the column names of the \link{StyledTable} object should be used. Be aware that the \link{StyledTable} object columns have the names \code{X1}, ..., \code{XN}, where \code{N} is the total number of columns. If you do not want to apply the change to all columns given in the \code{col_id} argument, but only to single cells which fullfill a special condition you can use also the variable name \code{X} which is substituted by each column name separately. The arguments \code{condition_text} and \code{condition} cannot be passed at the same time.}
}
\value{
The modified \link{StyledTable} object
}
\description{
Set vertical movement of cell content for LaTeX table generation
}
\seealso{
\code{\link[=set_excel_font_name]{set_excel_font_name()}}, \code{\link[=set_latex_font_name]{set_latex_font_name()}}, \code{\link[=set_excel_font_size]{set_excel_font_size()}}, \code{\link[=set_latex_font_size]{set_latex_font_size()}}, \code{\link[=set_font_color]{set_font_color()}}, \code{\link[=set_bold]{set_bold()}}, \code{\link[=set_excel_boldweight]{set_excel_boldweight()}}, \code{\link[=set_italic]{set_italic()}}, \code{\link[=set_underline]{set_underline()}}, \code{\link[=set_strikeout]{set_strikeout()}}, \code{\link[=set_rotation]{set_rotation()}}, \code{\link[=set_indent]{set_indent()}}, \code{\link[=set_fill_color]{set_fill_color()}}, \code{\link[=set_horizontal]{set_horizontal()}}, \code{\link[=set_excel_vertical]{set_excel_vertical()}}, \code{\link[=set_latex_vertical_move]{set_latex_vertical_move()}}, \code{\link[=set_latex_vertical_move]{set_latex_vertical_move()}}, \code{\link[=set_excel_row_height]{set_excel_row_height()}}, \code{\link[=set_latex_row_height]{set_latex_row_height()}}, \code{\link[=set_excel_col_width]{set_excel_col_width()}}, \code{\link[=set_latex_col_width]{set_latex_col_width()}}, \code{\link[=set_border_position]{set_border_position()}}, \code{\link[=set_border_color]{set_border_color()}}, \code{\link[=set_excel_border_pen]{set_excel_border_pen()}}, \code{\link[=set_excel_data_format]{set_excel_data_format()}}, \code{\link[=set_excel_pre_process]{set_excel_pre_process()}}, \code{\link[=set_latex_pre_process]{set_latex_pre_process()}}, \code{\link[=set_excel_wrapped]{set_excel_wrapped()}}, \code{\link[=set_excel_locked]{set_excel_locked()}}, \code{\link[=set_excel_hidden]{set_excel_hidden()}}
}
