% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHeightsWidths.R
\docType{methods}
\name{setExcelColWidths}
\alias{setExcelColWidths}
\alias{setExcelColWidths,StyledTable,numeric-method}
\title{Method setExcelColWidths.}
\usage{
setExcelColWidths(object, value, ...)

\S4method{setExcelColWidths}{StyledTable,numeric}(object, value, cols = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{value}{A numeric vector (length 1 or same lengt as \code{cols}) holding the column widths}

\item{...}{Various arguments}

\item{cols}{A vector of column numbers (N is substituted as \code{nCols(object)})}
}
\description{
Method setExcelColWidths.
}
