% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHeightsWidths.R
\name{setHeightsWidths}
\alias{setHeightsWidths}
\title{Function to set the rowHeights/latexRowHeights/colWidths/latexColWidths}
\usage{
setHeightsWidths(object, slotName, ind, value)
}
\arguments{
\item{object}{The styled table whoose heights/widths should be changed}

\item{slotName}{The name of the object slot that should be changed ("rowHeights", "latexRowHeights", "colWidths", "latexColWidths")}

\item{ind}{The row/col numbers that should be changed (subset of 1:N where N is the number of rows/cols of the styled table).}

\item{value}{A vector holding the row height/col width that should be applied}
}
\description{
Function to set the rowHeights/latexRowHeights/colWidths/latexColWidths
}
