% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHeightsWidths.R
\docType{methods}
\name{setLatexRowHeights}
\alias{setLatexRowHeights}
\alias{setLatexRowHeights,StyledTable,numeric-method}
\alias{setLatexRowHeiths,StyledTable,numeric-method}
\title{Method setLatexRowHeights.}
\usage{
setLatexRowHeights(object, value, ...)

\S4method{setLatexRowHeights}{StyledTable,numeric}(object, value, rows = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{value}{A numeric vector (length 1 or same lengt as \code{rows}) holding the row heights}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers (N is substituted as \code{nRows(object)})}
}
\description{
Method setLatexRowHeights.
}
