% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCellStyles.R
\docType{methods}
\name{setLatexVerticalMove}
\alias{setLatexVerticalMove}
\alias{setLatexVerticalMove,StyledTable-method}
\alias{setLatexVerticalMove,StyledTable,ANY-method}
\title{Set vertical movement of cell content for LaTeX table generation}
\usage{
setLatexVerticalMove(object, value, ...)

\S4method{setLatexVerticalMove}{StyledTable}(object, value, rows = NULL,
  cols = NULL, condition = NULL, conditionText = NULL)
}
\arguments{
\item{object}{A StyledTable}

\item{value}{The function that should be used for pre processing}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers to which the change should be applied to (\code{N} is substituted with the total number of table rows)}

\item{cols}{A vector of col numbers to which the change should be applied to (\code{N} is substituted with the total number of table cols)}

\item{condition}{An equation (non standard evaluation) that is evaluated on the data.frame \code{data} in order to decide on which rows or cells the change should be applied. In the equation the column names of the styled table should be used. Be aware that the styled table columns have the names \code{X1}, ..., \code{XN}, where \code{N} is the total number of columns. If you do not want to apply the change to all columns given in the \code{cols} argument, but only to single cells which fullfill a special condition you can use also the variable name \code{X} which is substituted by each column name separately. The arguments \code{conditionText} and \code{condition} cannot be passed at the same time.}

\item{conditionText}{An character string holding an equation that is evaluated on the data.frame \code{data} in order to decide on which rows or cells the change should be applied. In the equation the column names of the styled table should be used. Be aware that the styled table columns have the names \code{X1}, ..., \code{XN}, where \code{N} is the total number of columns. If you do not want to apply the change to all columns given in the \code{cols} argument, but only to single cells which fullfill a special condition you can use also the variable name \code{X} which is substituted by each column name separately. The arguments \code{conditionText} and \code{condition} cannot be passed at the same time.}
}
\description{
Set vertical movement of cell content for LaTeX table generation
}
