% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCellStyles.R
\docType{methods}
\name{setSpecificStyle}
\alias{setSpecificStyle}
\alias{setSpecificStyle,StyledTable,ANY,character,character-method}
\alias{setSpecificStyle,StyledTable,character,ANY-method}
\title{Set styles of multiple cells}
\usage{
setSpecificStyle(object, value, styleName, styleType, ...)

\S4method{setSpecificStyle}{StyledTable,ANY,character,character}(object,
  value, styleName, styleType, rows = NULL, cols = NULL,
  condition = NULL, conditionText = NULL, appendMode = "replace")
}
\arguments{
\item{object}{A StyledTable}

\item{value}{The value that should be set}

\item{styleName}{The name of the style slot that should be set}

\item{styleType}{The type which the value has to have (used for type checking)}

\item{...}{Various arguments}

\item{rows}{A vector of row numbers to which the change should be applied to (\code{N} is substituted with the total number of table rows)}

\item{cols}{A vector of col numbers to which the change should be applied to (\code{N} is substituted with the total number of table cols)}

\item{condition}{An equation (non standard evaluation) that is evaluated on the data.frame \code{data} in order to decide on which rows or cells the change should be applied. In the equation the column names of the styled table should be used. Be aware that the styled table columns have the names \code{X1}, ..., \code{XN}, where \code{N} is the total number of columns. If you do not want to apply the change to all columns given in the \code{cols} argument, but only to single cells which fullfill a special condition you can use also the variable name \code{X} which is substituted by each column name separately. The arguments \code{conditionText} and \code{condition} cannot be passed at the same time.}

\item{conditionText}{An character string holding an equation that is evaluated on the data.frame \code{data} in order to decide on which rows or cells the change should be applied. In the equation the column names of the styled table should be used. Be aware that the styled table columns have the names \code{X1}, ..., \code{XN}, where \code{N} is the total number of columns. If you do not want to apply the change to all columns given in the \code{cols} argument, but only to single cells which fullfill a special condition you can use also the variable name \code{X} which is substituted by each column name separately. The arguments \code{conditionText} and \code{condition} cannot be passed at the same time.}

\item{appendMode}{A character string that defines if the styling value of the cell should be replaced (appendMode = "replace"), if the new styling should be appended at the end of the current styling value of the cell (appendMode = "appendBehind"), if the new styling should be appended in front of the current styling value of the cell (appendMode = "appendBefore")}
}
\description{
Set styles of multiple cells
}
