% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{setSTCellStyle}
\alias{setSTCellStyle}
\alias{setSTCellStyle,STCellStyle,ANY,character-method}
\alias{setSTCellStyle,STCellStyle-method}
\alias{setSTCellStyle,MissingOrNull,ANY,character-method}
\title{Set a single attribute in a style class}
\usage{
setSTCellStyle(object, value, styleName)

\S4method{setSTCellStyle}{STCellStyle,ANY,character}(object, value,
  styleName)

\S4method{setSTCellStyle}{MissingOrNull,ANY,character}(object, value,
  styleName)
}
\arguments{
\item{object}{A STCellStyle object}

\item{value}{The value that should be set}

\item{styleName}{The name of the style setting that should be set}
}
\description{
Set a single attribute in a style class
}
