% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_cell.R
\name{setStyledCell}
\alias{setStyledCell}
\alias{setStyledCell,StyledCell,ANY,character-method}
\alias{setStyledCell,StyledCell-method}
\alias{setStyledCell,MissingOrNull,ANY,character-method}
\title{Set a single attribute in a style class}
\usage{
setStyledCell(sc, value, style_name)

\S4method{setStyledCell}{StyledCell,ANY,character}(sc, value, style_name)

\S4method{setStyledCell}{MissingOrNull,ANY,character}(sc, value, style_name)
}
\arguments{
\item{sc}{A StyledCell object}

\item{value}{The value that should be set}

\item{style_name}{The name of the style setting that should be set}
}
\description{
Set a single attribute in a style class
}
