% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{st_preamble}
\alias{st_preamble}
\title{Get the preamble code necessary to compile a stiled table}
\usage{
st_preamble()
}
\description{
Generate a string that can be included in the preamble of a latex document
to compile a styledtable object.
}
