% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{class}
\name{STCellStyle-class}
\alias{STCellStyle-class}
\alias{STCellStyle}
\alias{initialize,STCellStyle-method}
\title{Style class for single cells in styled tables}
\usage{
\S4method{initialize}{STCellStyle}(.Object, fontName, fontHeight,
  fontColor, isBold, isItalic, isStrikeout, underline, boldweight,
  isWrapped, horizontal, vertical, rotation, indent, borderPosition,
  borderColor, borderPen, foregroundColor, backgroundColor, fillPattern,
  dataFormat, isLocked, isHidden, latexVerticalMove, excelPreProcess,
  latexPreProcess, latexFontSize)
}
\arguments{
\item{.Object}{A STCellStyle object}

\item{fontName}{The name of the font that should be used}

\item{fontHeight}{The font height}

\item{fontColor}{The font color}

\item{isBold}{Is bold? TRUE/FALSE}

\item{isItalic}{Is italic? TRUE/FALSE}

\item{isStrikeout}{Is striked out? TRUE/FALSE}

\item{underline}{Numeric weight of the underline}

\item{boldweight}{Numeric font weight}

\item{isWrapped}{Is wrappend? TRUE/FALSE}

\item{horizontal}{String with horizontal alignment}

\item{vertical}{String with vertical alignment}

\item{rotation}{Numeric holding the rotation of the text}

\item{indent}{Numeric holding the indnentation}

\item{borderPosition}{Character vector with possible border positions}

\item{borderColor}{Color string (hex or color name)}

\item{borderPen}{String with border pen value}

\item{foregroundColor}{Color string for foreground color}

\item{backgroundColor}{Color string for background color}

\item{fillPattern}{String with fill pattern}

\item{dataFormat}{String for Excel-Cell-Data-Format}

\item{isLocked}{Is Locked? TRUE/FALSE}

\item{isHidden}{Is Hidden? TRUE/FALSE}

\item{latexVerticalMove}{A latex string that gives the distance for vertically moving the cell content (e.g.: in order to to vertically center it).}

\item{excelPreProcess}{A function that can be used to pre process cell values for the Excel table generation}

\item{latexPreProcess}{A function that can be used to pre process cell values for the LaTeX table generation}

\item{latexFontSize}{The LaTeX command to set the font size}
}
\description{
This class holds all style settings for a single excel cell. The values are
filled in by the user.
}
