% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stringify}
\alias{stringify}
\title{Stringify vector}
\usage{
stringify(
  x,
  before = NULL,
  after = NULL,
  collapse = ", ",
  quote = "'",
  n_max = 10,
  further = "..."
)
}
\arguments{
\item{x}{An object, which should be turned into a string. Usually an atomic
object (a vector).}

\item{before}{A string placed before each vector entry.}

\item{after}{A string placed after each vector entry.}

\item{collapse}{A string, used for separating each vector entry.}

\item{quote}{A string, holding  the symbol which should be used for quoting
every entry of \code{x}. Default is \verb{'} which gives the result \code{'entry1'}.}

\item{n_max}{A positive integer or \code{Inf}, defining the maximum number of
displayed vector elements. All further elements are displayed with the
text defined in \code{further}.}

\item{further}{A string indicating further (after \code{n_max} displayed elements)
vector elements.}
}
\value{
A string showing the entries of the vector.
}
\description{
Turn a vector (e.g. \code{c(1, 2, 3)}) into a string (e.g. \code{"'1', '2', '3'"}).
This function is particularly useful when creating error messages,
describing problematic values.
}
