% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\docType{methods}
\name{setStyle}
\alias{setStyle}
\alias{setStyle,Style,ANY,character-method}
\alias{setStyle,Style-method}
\alias{setStyle,MissingOrNull,ANY,character-method}
\alias{setStyle,Style-method}
\title{Set a single attribute in a style class}
\usage{
setStyle(object, value, styleName)

\S4method{setStyle}{Style,ANY,character}(object, value, styleName)

\S4method{setStyle}{MissingOrNull,ANY,character}(object, value, styleName)
}
\arguments{
\item{object}{A Style object}

\item{value}{The value that should be set}

\item{styleName}{The name of the style setting that should be set}
}
\description{
Set a single attribute in a style class
}
