% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_cell.R
\docType{class}
\name{StyledCell-class}
\alias{StyledCell-class}
\alias{StyledCell}
\alias{initialize,StyledCell-method}
\title{Style class for single cells in \link{StyledTable} objects}
\usage{
\S4method{initialize}{StyledCell}(
  .Object,
  excel_font_name,
  latex_font_name,
  excel_font_size,
  font_color,
  bold,
  italic,
  strikeout,
  underline,
  excel_boldweight,
  excel_wrapped,
  horizontal,
  vertical,
  rotation,
  indent,
  border_position,
  border_color,
  excel_border_pen,
  fill_color,
  excel_background_color,
  excel_fill_pattern,
  excel_data_format,
  excel_locked,
  excel_hidden,
  latex_vertical_move,
  excel_pre_process,
  latex_pre_process,
  html_pre_process,
  latex_font_size,
  html_class,
  html_id
)
}
\arguments{
\item{.Object}{A StyledCell object}

\item{excel_font_name}{The name of the font that should be used for the Excel table generation}

\item{latex_font_name}{The name of the font that should be used for the LaTeX table generation}

\item{excel_font_size}{The font height}

\item{font_color}{The font color}

\item{bold}{Is bold? TRUE/FALSE}

\item{italic}{Is italic? TRUE/FALSE}

\item{strikeout}{Is striked out? TRUE/FALSE}

\item{underline}{Numeric weight of the underline}

\item{excel_boldweight}{Numeric font weight}

\item{excel_wrapped}{Is wrappend? TRUE/FALSE}

\item{horizontal}{String with horizontal alignment}

\item{vertical}{String with vertical alignment}

\item{rotation}{Numeric holding the rotation of the text}

\item{indent}{Numeric holding the indnentation}

\item{border_position}{Character vector with possible border positions}

\item{border_color}{Color string (hex or color name)}

\item{excel_border_pen}{String with border pen value}

\item{fill_color}{Color string for foreground color}

\item{excel_background_color}{Color string for background color}

\item{excel_fill_pattern}{String with fill pattern}

\item{excel_data_format}{String for Excel-Cell-Data-Format}

\item{excel_locked}{Is Locked? TRUE/FALSE}

\item{excel_hidden}{Is Hidden? TRUE/FALSE}

\item{latex_vertical_move}{A latex string that gives the distance for vertically moving the cell content (e.g.: in order to to vertically center it).}

\item{excel_pre_process}{A function that can be used to pre process cell values for the Excel table generation}

\item{latex_pre_process}{A function that can be used to pre process cell values for the LaTeX table generation}

\item{html_pre_process}{A function that can be used to pre process cell values for the HTML table generation}

\item{latex_font_size}{The LaTeX command to set the font size}

\item{html_class}{A character vector holding the html class names assigned to the table cell (\verb{<td>} dom element)}

\item{html_id}{A class vector holding html ids, which will be added to the table cells (\verb{<td>} dom element)}
}
\description{
This class holds all style settings for a single excel cell. The values are
filled in by the user.
}
