% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyledTable.R
\docType{class}
\name{StyledTable-class}
\alias{StyledTable-class}
\alias{StyledTable}
\alias{initialize,StyledTable-method}
\title{Class StyledTable}
\usage{
\S4method{initialize}{StyledTable}(.Object, ...)
}
\arguments{
\item{.Object}{A StyledTable}

\item{...}{All slots that should be set}
}
\description{
Class StyledTable

Constructor method of StyledTable Class.
}
