% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_table.R
\docType{methods}
\name{count_cols}
\alias{count_cols}
\alias{count_cols,StyledTable-method}
\title{Count number of columns in a \link{StyledTable}}
\usage{
count_cols(st)

\S4method{count_cols}{StyledTable}(st)
}
\arguments{
\item{st}{A \link{StyledTable} object}
}
\description{
Count number of columns in a \link{StyledTable}
}
