% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_table.R
\name{count_rows}
\alias{count_rows}
\alias{count_rows,StyledTable-method}
\title{Count number of rows in a \link{StyledTable}}
\usage{
count_rows(st)

\S4method{count_rows}{StyledTable}(st)
}
\arguments{
\item{st}{A \link{StyledTable} object}
}
\description{
Count number of rows in a \link{StyledTable}
}
