% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyledTable.R
\docType{methods}
\name{nCols}
\alias{nCols}
\alias{nCols,StyledTable-method}
\title{Number of cols in a StyledTable}
\usage{
nCols(object)

\S4method{nCols}{StyledTable}(object)
}
\arguments{
\item{object}{A styled table}
}
\description{
Number of cols in a StyledTable
}
