% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyledTable.R
\docType{methods}
\name{nRows}
\alias{nRows}
\alias{nRows,StyledTable-method}
\title{Number of rows in a StyledTable}
\usage{
nRows(object)

\S4method{nRows}{StyledTable}(object)
}
\arguments{
\item{object}{A styled table}
}
\description{
Number of rows in a StyledTable
}
