% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyledTable.R
\docType{methods}
\name{setStyles}
\alias{setStyles}
\alias{setStyles,StyledTable-method}
\title{Set styles of StyledTable cells}
\usage{
setStyles(object, ...)

\S4method{setStyles}{StyledTable}(object, value, styleName = "",
  rows = NULL, cols = NULL, appendMode = "replace")
}
\arguments{
\item{object}{A StyledTable}

\item{...}{Various Arguments}

\item{value}{The value that should be set}

\item{styleName}{The name of the style slot that should be set}

\item{rows}{A vector of row numbers to which the change should be applied to}

\item{cols}{A vector of col numbers to which the change should be applied to}

\item{appendMode}{A character string that defines if the styling value of the cell should be replaced (appendMode = "replace"), if the new styling should be appended at the end of the current styling value of the cell (appendMode = "appendBehind"), if the new styling should be appended in front of the current styling value of the cell (appendMode = "appendBefore")}
}
\description{
Set styles of StyledTable cells
}
