% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_table.R
\name{st_bind_cols}
\alias{st_bind_cols}
\alias{st_bind_cols,ANY-method}
\alias{st_bind_cols-method}
\title{Create a \link{StyledTable} object}
\usage{
st_bind_cols(..., keep_header = FALSE)

\S4method{st_bind_cols}{ANY}(..., keep_header = FALSE)
}
\arguments{
\item{...}{Multiple matrices, data.frames or \link{StyledTable} objects that should be concatenated}

\item{keep_header}{(optional) A logical flag, if the column names of passed
\code{data.frame} should be written in the first line of the resulting
\link{StyledTable} object object.}
}
\value{
A \link{StyledTable} object object
}
\description{
This function creates a \link{StyledTable} object from a single data.frame or multiple data.frames or other \link{StyledTable} objects. If multiple \link{StyledTable} objects or data.frames are supplied, then they are concatenated vertically. Therefore, all supplied data.frames and \link{StyledTable} objects must have the same number of columns.
}
