% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_table.R
\name{styled_table}
\alias{styled_table}
\alias{styled_table,ANY-method}
\alias{styled_table-method}
\title{Create a \link{StyledTable} object}
\usage{
styled_table(..., keep_header = FALSE, nrow = NULL)

\S4method{styled_table}{ANY}(..., keep_header = FALSE, nrow = NULL)
}
\arguments{
\item{...}{Multiple matrices, data.frames or \link{StyledTable} objects that should be concatenated}

\item{keep_header}{(optional) A logical flag, if the column names of passed
\code{data.frame} should be written in the first line of the resulting
\link{StyledTable} object object.}

\item{nrow}{(optional) If the passed object is an atomic vector, then the
it can be transformed into a \code{\link[=matrix]{matrix()}} by defining the number of rows it
should be split into.}
}
\value{
A \link{StyledTable} object object
}
\description{
This function creates a \link{StyledTable} object from a single data.frame or multiple data.frames or other \link{StyledTable} objects. If multiple \link{StyledTable} objects or data.frames are supplied, then they are concatenated vertically. Therefore, all supplied data.frames and \link{StyledTable} objects must have the same number of columns.
}
