% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyledTable.R
\docType{methods}
\name{styledTable}
\alias{styledTable}
\alias{styledTable,ANY-method}
\alias{styledTable-method}
\title{Create a styled table}
\usage{
styledTable(...)

\S4method{styledTable}{ANY}(...)
}
\arguments{
\item{...}{Multiple matrices, data.frames or styled tables that should be concatenated}
}
\value{
A styled table object
}
\description{
This function creates a styled table from a single data.frame or multiple data.frames and other styled tables. If multiple styled tables and data.frames are supplied, then they are concatenated vertically. Therefore, all supplied data.frames and styled tables must have the same number of columns.
}
