% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{substitute_col_id}
\alias{substitute_col_id}
\title{Substitute \code{col_id} and \code{row_id}}
\usage{
substitute_col_id(st, col_id, stack_level = 1, error_handler)
}
\arguments{
\item{st}{The \link{StyledTable} object}

\item{col_id}{The passed expression that should be evaluated to the column ids}

\item{stack_level}{The level number of the environment stack (1: parent.frame, 2: parent.parent.frame))}

\item{error_handler}{Error handling function. Takes a message string}
}
\value{
The evaluated column ids (unbroken numeric vector)
}
\description{
This function is used to substitute \code{n_col} and \code{n_row} inside
of the expressions passed into \code{col_id} and \code{row_id}
}
