% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{substitute_row_id}
\alias{substitute_row_id}
\title{Substitute \code{row_id}}
\usage{
substitute_row_id(st, row_id, stack_level = 1, error_handler)
}
\arguments{
\item{st}{The \link{StyledTable} object}

\item{row_id}{The passed expression that should be evaluated to the row ids}

\item{stack_level}{The level number of the environment stack (1: parent.frame, 2: parent.parent.frame))}

\item{error_handler}{Error handling function. Takes a message string}
}
\value{
The evaluated row ids (unbroken numeric vector)
}
\description{
This function is used to substitute \code{n_row} inside
of the expression passed into \code{row_id}
}
