% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{transform_char}
\alias{transform_char}
\title{Transform text}
\usage{
transform_char(var = NULL, pre = "", post = "", toUpper = TRUE)
}
\arguments{
\item{var}{(string) Text that should be transformed}

\item{pre}{(string) Text that should be appended in front}

\item{post}{(string) Text that should be appended at the end}

\item{toUpper}{(logical) Should the text be transformed to upper case?}
}
\description{
Transform a text to upper case and maybe append some text at front of the
beginning or at the end
}
