% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.pdf.R
\name{write.pdf}
\alias{write.pdf}
\title{Write a styledTable object to pdf}
\usage{
write.pdf(x, file = "table.pdf", resize_pdf = TRUE)
}
\arguments{
\item{x}{a styledTable object to save.}

\item{file}{the path to the output file.}

\item{resize_pdf}{Should the pdf be resized to the dimensions of
the table? Default is \code{TRUE}.}
}
\description{
Use \link[tools:texi2pdf]{tools::texi2pdf} to compile the output from \link{createLatexTable}
to a pdf. This function works similat ro \link{write.csv} in the sense
that the first argument is the table to write and the second
argument is a local or global path.
}
