% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\docType{methods}
\name{write_excel}
\alias{write_excel}
\alias{write_excel,ANY,StyledTable-method}
\title{Write a StyledTable to an Excel sheet}
\usage{
write_excel(sheet, st, ...)

\S4method{write_excel}{ANY,StyledTable}(sheet, st, first_row = NULL,
  first_col = NULL)
}
\arguments{
\item{sheet}{An xlsx sheet object}

\item{st}{A \link{StyledTable} object object}

\item{...}{Various arguments}

\item{first_row}{The number of the starting row of the table in the sheet}

\item{first_col}{The number of the starting col of the table in the sheet}
}
\description{
Write a StyledTable to an Excel sheet
}
\seealso{
\code{\link[=write_pdf]{write_pdf()}}, \code{\link[=write_png]{write_png()}}, \code{\link[=create_latex_table]{create_latex_table()}}, \code{\link[=append_latex_table]{append_latex_table()}}, \code{\link[=create_latex_table_body]{create_latex_table_body()}}
}
