% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pdf.R
\name{write_pdf}
\alias{write_pdf}
\title{Write a styledTable object to pdf}
\usage{
write_pdf(st, file = "table.pdf", resize = TRUE)
}
\arguments{
\item{st}{The styledTable object to save.}

\item{file}{The path where the output file should be saved.}

\item{resize}{Should the pdf be resized to the dimensions of
the table? Default is \code{TRUE}.}
}
\description{
Use \code{\link[tools:texi2dvi]{tools::texi2pdf()}} to compile the output from
\code{\link[=create_latex_table]{create_latex_table()}} to a pdf and save the file.
}
\examples{
\dontrun{
library(dplyr)
mtcars[1:5, 1:2] \%>\%
  styled_table \%>\%
  write_pdf("table_image.pdf")
}
}
\seealso{
\code{\link[=write_png]{write_png()}}, \code{\link[=write_excel]{write_excel()}}, \code{\link[=create_latex_table]{create_latex_table()}}, \code{\link[=append_latex_table]{append_latex_table()}}, \code{\link[=create_latex_table_body]{create_latex_table_body()}}
}
