% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_png.R
\name{write_png}
\alias{write_png}
\title{Write a styledTable object to png}
\usage{
write_png(st, file = "table.png", resize = TRUE)
}
\arguments{
\item{st}{a styledTable object to save.}

\item{file}{the path to the output file.}

\item{resize}{Should the Png be resized to the dimensions of
the table? Default is \code{TRUE}.}
}
\description{
Use \code{\link[tools:texi2pdf]{tools::texi2pdf()}} to compile the output from
\code{\link[=create_latex_table]{create_latex_table()}} to a png and save the file.
}
\examples{
library(dplyr)
mtcars[1:5, 1:2] \%>\%
  styled_table \%>\%
  write_png("table_image.png")
}
\seealso{
\code{\link[=write_pdf]{write_pdf()}}, \code{\link[=write_excel]{write_excel()}}, \code{\link[=create_latex_table]{create_latex_table()}}, \code{\link[=append_latex_table]{append_latex_table()}}, \code{\link[=create_latex_table_body]{create_latex_table_body()}}
}
