% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePng.R
\name{writePng}
\alias{writePng}
\title{Write a styledTable object to png}
\usage{
writePng(x, file = "table.png", resizePng = TRUE)
}
\arguments{
\item{x}{a styledTable object to save.}

\item{file}{the path to the output file.}

\item{resizePng}{Should the Png be resized to the dimensions of
the table? Default is \code{TRUE}.}
}
\description{
Use \code{tools::texi2dvi} and \code{ghostconvert} to compile the output from \code{createLatexTable}
to a pdf and then into a png. This function works similat to \code{write.csv} in the sense
that the first argument is the table to write and the second
argument is a local or global path.
}
