% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToExcelSheet.R
\docType{methods}
\name{writeToExcelSheet}
\alias{writeToExcelSheet}
\alias{writeToExcelSheet,ANY,StyledTable-method}
\title{Write a StyledTable to an Excel sheet}
\usage{
writeToExcelSheet(sheet, object, ...)

\S4method{writeToExcelSheet}{ANY,StyledTable}(sheet, object, firstRow = NULL,
  firstCol = NULL)
}
\arguments{
\item{sheet}{An xlsx sheet object}

\item{object}{A StyledTable}

\item{...}{Various arguments}

\item{firstRow}{The number of the starting row of the table in the sheet}

\item{firstCol}{The number of the starting col of the table in the sheet}
}
\description{
Write a StyledTable to an Excel sheet
}
