% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{create_html_tag_style}
\alias{create_html_tag_style}
\title{Create a string defining the style property of an html tag}
\usage{
create_html_tag_style(values = NULL)
}
\arguments{
\item{values}{An optional named list or named character vector holding a set
of styling informations for an html dom element.}
}
\value{
A string holding the text for the html tag styling property
}
\description{
Create a string defining the style property of an html tag
}
