% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_cell.R
\docType{methods}
\name{getXlsxFontCellStyle}
\alias{getXlsxFontCellStyle}
\alias{getXlsxFontCellStyle,ANY,StyledCell-method}
\alias{getXlsxFontCellStyle,StyledCell-method}
\title{Method getXlsxFontCellStyle}
\usage{
getXlsxFontCellStyle(wb, sc)

\S4method{getXlsxFontCellStyle}{ANY,StyledCell}(wb, sc)
}
\arguments{
\item{wb}{An xlsx workbook object}

\item{sc}{A StyledCell object}
}
\description{
Method getXlsxFontCellStyle
}
