% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_table.R
\docType{class}
\name{StyledTable-class}
\alias{StyledTable-class}
\alias{StyledTable}
\alias{initialize,StyledTable-method}
\title{Class StyledTable}
\usage{
\S4method{initialize}{StyledTable}(.Object, ...)
}
\arguments{
\item{.Object}{A StyledTable}

\item{...}{All slots that should be set}
}
\description{
An \code{S4 class} which is used to export your data as a beautifully styled table.
Convert your \code{data.frame} or \code{matrix} into a \code{StyledTable} object by using the \code{\link[=styled_table]{styled_table()}} function.
}
\seealso{
\code{\link[=styled_table]{styled_table()}}

\code{\link[=write_pdf]{write_pdf()}}, \code{\link[=write_png]{write_png()}}, \code{\link[=write_excel]{write_excel()}}, \code{\link[=create_latex_table]{create_latex_table()}}, \code{\link[=append_latex_table]{append_latex_table()}}, \code{\link[=create_latex_table_body]{create_latex_table_body()}}

\code{\link[=set_excel_font_name]{set_excel_font_name()}}, \code{\link[=set_latex_font_name]{set_latex_font_name()}}, \code{\link[=set_excel_font_size]{set_excel_font_size()}}, \code{\link[=set_latex_font_size]{set_latex_font_size()}}, \code{\link[=set_font_color]{set_font_color()}}, \code{\link[=set_bold]{set_bold()}}, \code{\link[=set_excel_boldweight]{set_excel_boldweight()}}, \code{\link[=set_italic]{set_italic()}}, \code{\link[=set_underline]{set_underline()}}, \code{\link[=set_strikeout]{set_strikeout()}}, \code{\link[=set_rotation]{set_rotation()}}, \code{\link[=set_indent]{set_indent()}}, \code{\link[=set_fill_color]{set_fill_color()}}, \code{\link[=set_horizontal]{set_horizontal()}}, \code{\link[=set_excel_vertical]{set_excel_vertical()}}, \code{\link[=set_latex_vertical_move]{set_latex_vertical_move()}}, \code{\link[=set_latex_vertical_move]{set_latex_vertical_move()}}, \code{\link[=set_excel_row_height]{set_excel_row_height()}}, \code{\link[=set_latex_row_height]{set_latex_row_height()}}, \code{\link[=set_excel_col_width]{set_excel_col_width()}}, \code{\link[=set_latex_col_width]{set_latex_col_width()}}, \code{\link[=set_border_position]{set_border_position()}}, \code{\link[=set_border_color]{set_border_color()}}, \code{\link[=set_excel_border_pen]{set_excel_border_pen()}}, \code{\link[=set_excel_data_format]{set_excel_data_format()}}, \code{\link[=set_excel_pre_process]{set_excel_pre_process()}}, \code{\link[=set_latex_pre_process]{set_latex_pre_process()}}, \code{\link[=set_excel_wrapped]{set_excel_wrapped()}}, \code{\link[=set_excel_locked]{set_excel_locked()}}, \code{\link[=set_excel_hidden]{set_excel_hidden()}}

\code{\link[=remove_col]{remove_col()}}, \code{\link[=remove_row]{remove_row()}}

\code{\link[=merge_cells]{merge_cells()}}, \code{\link[=merge_equal_cells]{merge_equal_cells()}}

\code{\link[=format_stat_header]{format_stat_header()}}, \code{\link[=format_stat_body]{format_stat_body()}}, \code{\link[=format_stat_sub_heading]{format_stat_sub_heading()}}, \code{\link[=format_stat_absolute]{format_stat_absolute()}}, \code{\link[=format_stat_relative]{format_stat_relative()}}

\code{\link[=create_cross_table_body]{create_cross_table_body()}}, \code{\link[=create_cross_table_header]{create_cross_table_header()}}

\code{\link[=count_rows]{count_rows()}}, \code{\link[=count_cols]{count_cols()}}
}
