% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_table.R
\name{setStyles}
\alias{setStyles}
\alias{setStyles,StyledTable-method}
\title{Set styles of StyledTable cells}
\usage{
setStyles(st, ...)

\S4method{setStyles}{StyledTable}(
  st,
  value,
  style_name = "",
  row_id = NULL,
  col_id = NULL,
  append_mode = "replace"
)
}
\arguments{
\item{st}{A \link{StyledTable} object}

\item{...}{Various Arguments}

\item{value}{The value that should be set}

\item{style_name}{The name of the style slot that should be set}

\item{row_id}{A vector of row numbers to which the change should be applied to}

\item{col_id}{A vector of col numbers to which the change should be applied to}

\item{append_mode}{A character string that defines if the styling value of the cell should be replaced (append_mode = "replace"), if the new styling should be appended at the end of the current styling value of the cell (append_mode = "appendBehind"), if the new styling should be appended in front of the current styling value of the cell (append_mode = "appendBefore")}
}
\description{
Set styles of StyledTable cells
}
